requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.3
llama-index==0.9.48
crewai==0.1.25

# LLM Providers
openai==1.10.0
cohere==4.47
anthropic==0.8.1

# Vector Databases
chromadb==0.4.22
faiss-cpu==1.7.4
neo4j==5.16.0

# Embeddings
sentence-transformers==2.3.1
transformers==4.37.0
torch==2.1.2

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
pdf2image==1.16.3

# Audio/Video Processing
whisper==1.1.10
moviepy==1.0.3
youtube-transcript-api==0.6.2

# Search & Retrieval
rank-bm25==0.2.2
pymilvus==2.3.5
duckduckgo-search==4.1.1

# Evaluation & Metrics
ragas==0.1.1
datasets==2.16.1
nltk==3.8.1
rouge-score==0.1.2
bert-score==0.3.13

# Web Framework & API
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0
plotly==5.18.0
pydantic==2.5.3
python-multipart==0.0.6

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1

# MLOps & Monitoring
mlflow==2.10.0
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Security & Guardrails
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1
tqdm==4.66.1
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Graph Processing
networkx==3.2.1
pyvis==0.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Jupyter & Notebooks
jupyter==1.0.0
ipywidgets==8.1.1
notebook==7.0.6

# OCR Dependencies
easyocr==1.7.1

# Additional NLP
spacy-transformers==1.3.4
textstat==0.7.3

# Rate Limiting
slowapi==0.1.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration Management
hydra-core==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
umap-learn==0.5.5

# Async Support
asyncio==3.4.3
aiofiles==23.2.1
